+++ rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/annotation/RocketMQMessageListenerBeanPostProcessor.java
@@ -1,6 +1,22 @@
-package org.apache.rocketmq.client.client.annotation;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.rocketmq.client.annotation;
 
-import org.apache.rocketmq.client.client.autoconfigure.ListenerContainerConfiguration;
+import org.apache.rocketmq.client.autoconfigure.ListenerContainerConfiguration;
 import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.InitializingBean;
@@ -16,9 +32,6 @@ import java.util.Map;
 import java.util.function.BiFunction;
 import java.util.stream.Collectors;
 
-/**
- * @author Akai
- */
 public class RocketMQMessageListenerBeanPostProcessor implements ApplicationContextAware, BeanPostProcessor, InitializingBean {
 
     private ApplicationContext applicationContext;
@@ -67,7 +80,6 @@ public class RocketMQMessageListenerBeanPostProcessor implements ApplicationCont
                         .collect(Collectors.toList());
                 this.enhancer = (attrs, element) -> {
                     Map<String, Object> newAttrs = attrs;
-                    //遍历所有注解增强器，将前一次注解增强得到的结果作为下一个注解增强器的参数，最后返回多个注解增强器处理后的结果
                     for (AnnotationEnhancer enh : enhancers) {
                         newAttrs = enh.apply(newAttrs, element);
                     }
