+++ rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/autoconfigure/ExtTemplateResetConfiguration.java
@@ -1,13 +1,29 @@
-package org.apache.rocketmq.client.client.autoconfigure;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.rocketmq.client.autoconfigure;
 
-import org.apache.rocketmq.client.client.annotation.ExtProducerResetConfiguration;
-import org.apache.rocketmq.client.client.support.RocketMQMessageConverter;
-import org.apache.rocketmq.client.client.support.RocketMQUtil;
+import org.apache.rocketmq.client.annotation.ExtProducerResetConfiguration;
+import org.apache.rocketmq.client.support.RocketMQMessageConverter;
+import org.apache.rocketmq.client.support.RocketMQUtil;
 import org.apache.rocketmq.client.apis.ClientConfiguration;
 import org.apache.rocketmq.client.apis.ClientServiceProvider;
 import org.apache.rocketmq.client.apis.producer.ProducerBuilder;
 
-import org.apache.rocketmq.client.client.core.RocketMQClientTemplate;
+import org.apache.rocketmq.client.core.RocketMQClientTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.aop.framework.AopProxyUtils;
@@ -27,9 +43,6 @@ import java.time.Duration;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-/**
- * @author Akai
- */
 @Configuration
 public class ExtTemplateResetConfiguration implements ApplicationContextAware, SmartInitializingSingleton {
