+++ rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/autoconfigure/ListenerContainerConfiguration.java
@@ -1,9 +1,25 @@
-package org.apache.rocketmq.client.client.autoconfigure;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.rocketmq.client.autoconfigure;
 
-import org.apache.rocketmq.client.client.annotation.RocketMQMessageListener;
-import org.apache.rocketmq.client.client.core.RocketMQListener;
-import org.apache.rocketmq.client.client.support.DefaultListenerContainer;
-import org.apache.rocketmq.client.client.support.RocketMQMessageConverter;
+import org.apache.rocketmq.client.annotation.RocketMQMessageListener;
+import org.apache.rocketmq.client.core.RocketMQListener;
+import org.apache.rocketmq.client.support.DefaultListenerContainer;
+import org.apache.rocketmq.client.support.RocketMQMessageConverter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.BeansException;
@@ -18,9 +34,6 @@ import org.springframework.util.Assert;
 import java.time.Duration;
 import java.util.concurrent.atomic.AtomicLong;
 
-/**
- * @author Akai
- */
 @Configuration
 public class ListenerContainerConfiguration implements ApplicationContextAware {
     private final static Logger log = LoggerFactory.getLogger(ListenerContainerConfiguration.class);
@@ -52,8 +65,7 @@ public class ListenerContainerConfiguration implements ApplicationContextAware {
         String containerBeanName = String.format("%s_%s", DefaultListenerContainer.class.getName(),
                 counter.incrementAndGet());
         GenericApplicationContext genericApplicationContext = (GenericApplicationContext) applicationContext;
-        genericApplicationContext.registerBean(containerBeanName, DefaultListenerContainer.class,
-                () -> createRocketMQListenerContainer(containerBeanName, bean, annotation));
+        genericApplicationContext.registerBean(containerBeanName, DefaultListenerContainer.class, () -> createRocketMQListenerContainer(containerBeanName, bean, annotation));
         DefaultListenerContainer container = genericApplicationContext.getBean(containerBeanName,
                 DefaultListenerContainer.class);
         if (!container.isRunning()) {
