+++ rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/autoconfigure/RocketMQTransactionConfiguration.java
@@ -1,8 +1,24 @@
-package org.apache.rocketmq.client.client.autoconfigure;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.rocketmq.client.autoconfigure;
 
-import org.apache.rocketmq.client.client.annotation.RocketMQTransactionListener;
+import org.apache.rocketmq.client.annotation.RocketMQTransactionListener;
 import org.apache.rocketmq.client.apis.producer.TransactionChecker;
-import org.apache.rocketmq.client.client.core.RocketMQClientTemplate;
+import org.apache.rocketmq.client.core.RocketMQClientTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.aop.framework.AopProxyUtils;
@@ -18,9 +34,6 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-/**
- * @author Akai
- */
 @Configuration
 public class RocketMQTransactionConfiguration implements ApplicationContextAware, SmartInitializingSingleton {
     private final static Logger log = LoggerFactory.getLogger(RocketMQTransactionConfiguration.class);
@@ -32,7 +45,6 @@ public class RocketMQTransactionConfiguration implements ApplicationContextAware
         this.applicationContext = (ConfigurableApplicationContext) applicationContext;
     }
 
-    //获取被@RocketMQTransactionListener标记的类
     @Override
     public void afterSingletonsInstantiated() {
         Map<String, Object> beans = this.applicationContext.getBeansWithAnnotation(RocketMQTransactionListener.class)
@@ -50,7 +62,6 @@ public class RocketMQTransactionConfiguration implements ApplicationContextAware
         if (Objects.isNull(annotation)) {
             throw new IllegalStateException("The transactionListener annotation is missing");
         }
-        //获取注解上的template,默认为RocketMQGRpcTemplate
         RocketMQClientTemplate rocketMQTemplate = (RocketMQClientTemplate) applicationContext.getBean(annotation.rocketMQTemplateBeanName());
         if ((rocketMQTemplate.getProducerBuilder()) != null) {
             rocketMQTemplate.getProducerBuilder().setTransactionChecker((TransactionChecker) bean);
