--- rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/client/annotation/RocketMQTransactionListener.java
@@ -1,16 +1,0 @@
-package org.apache.rocketmq.client.client.annotation;
-
-import org.springframework.stereotype.Component;
-
-import java.lang.annotation.*;
-
-/**
- * @author Akai
- */
-@Target({ElementType.TYPE, ElementType.ANNOTATION_TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Component
-public @interface RocketMQTransactionListener {
-    String rocketMQTemplateBeanName() default "rocketMQClientTemplate";
-}
