--- rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/client/core/RocketMQTransactionChecker.java
@@ -1,12 +1,0 @@
-package org.apache.rocketmq.client.client.core;
-
-import org.apache.rocketmq.client.apis.message.MessageView;
-import org.apache.rocketmq.client.apis.producer.TransactionChecker;
-import org.apache.rocketmq.client.apis.producer.TransactionResolution;
-
-/**
- * @author Akai
- */
-public interface RocketMQTransactionChecker extends TransactionChecker {
-    TransactionResolution check(MessageView var1);
-}
