--- rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/client/support/RocketMQListenerContainer.java
@@ -1,9 +1,0 @@
-package org.apache.rocketmq.client.client.support;
-
-import org.springframework.beans.factory.DisposableBean;
-
-/**
- * @author Akai
- */
-public interface RocketMQListenerContainer extends DisposableBean {
-}
