+++ rocketmq-client-spring-boot/src/main/java/org/apache/rocketmq/client/support/RocketMQUtil.java
@@ -1,7 +1,28 @@
-package org.apache.rocketmq.client.client.support;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.rocketmq.client.support;
 
-import org.apache.rocketmq.client.client.autoconfigure.RocketMQProperties;
-import org.apache.rocketmq.client.apis.*;
+import org.apache.rocketmq.client.apis.ClientConfiguration;
+import org.apache.rocketmq.client.apis.ClientServiceProvider;
+import org.apache.rocketmq.client.apis.SessionCredentialsProvider;
+import org.apache.rocketmq.client.apis.StaticSessionCredentialsProvider;
+import org.apache.rocketmq.client.apis.ClientConfigurationBuilder;
+import org.apache.rocketmq.client.autoconfigure.RocketMQProperties;
+
 import org.apache.rocketmq.client.apis.consumer.FilterExpression;
 import org.apache.rocketmq.client.apis.consumer.FilterExpressionType;
 import org.slf4j.Logger;
@@ -15,9 +36,6 @@ import java.nio.charset.Charset;
 import java.time.Duration;
 import java.util.Objects;
 
-/**
- * @author Akai
- */
 public class RocketMQUtil {
 
     private static final Logger log = LoggerFactory.getLogger(RocketMQUtil.class);
@@ -146,7 +164,7 @@ public class RocketMQUtil {
             log.info("do not support your filterExpressionType {}", type);
         }
         FilterExpressionType filterExpressionType = "tag".equalsIgnoreCase(type) ? FilterExpressionType.TAG : FilterExpressionType.SQL92;
-        FilterExpression filterExpression= new FilterExpression(tag, filterExpressionType);
+        FilterExpression filterExpression = new FilterExpression(tag, filterExpressionType);
         return filterExpression;
     }
 }
